Added script-based test options
A new section has been added to the README.md file explaining how to set test options directly in the script. This includes an example of setting virtual users (vus) and duration within a JavaScript code block. The aim is to provide an alternative to manually typing these options each time, enhancing usability and efficiency.