Added new API check script
A new script has been added to perform checks on the API. This includes making a GET request to the server and logging the response. It also verifies that the response status is 2xx, the body of the response is not empty, and that the duration of the response is less than 500ms.